requirements.txt
Flask
line-bot-sdk<3.0.0
python-dotenv
gunicorn